requirements.txt
requests
python-socketio[asyncio]
fastapi
uvicorn
loguru
